backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiobotocore>=3.0.0",
    "aiosmtplib>=5.0.0",
    "aiosqlite>=0.22.1",
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "bcrypt==4.0.1",
    "cryptography>=46.0.3",
    "fastapi>=0.126.0",
    "fastapi-cache2>=0.2.2",
    "httpx>=0.28.1",
    "passlib>=1.7.4",
    "pillow>=12.0.0",
    "psycopg2>=2.9.11",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.12.5",
    "pyjwt>=2.10.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
    "python-multipart>=0.0.21",
    "redis>=7.1.0",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.38.0",
    "websockets>=16.0",
]
